refactor: 예약 가능 시간 조회 책임 분리
